Make it possible to check only cursor column or line
Some of the exercises fail because the start position and instructions do not
result in the cursor coordinates expected. An example of this is 3.2.1 ('gg')
where the cursor starts in (15,5) and the expectation of (1,1) cannot be met
with just `gg`.

However the test is too strict, we only want to know that the cursor is in the
first row, the column isn't important.